requirements.txt
PySide6>=6.7,<7
QtAwesome==1.4.0
pyqtdarktheme==2.1.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
bcrypt>=4.1.2
python-dotenv>=1.0
reportlab>=4.1
pytest>=8.0  
matplotlib==3.9.2